wip: improve shadow depth resolution with depth-clamp
1) this adds code to emulate depth-clamp. Side
effect disables early-z culling for the shadow
pass. how much of a problem is it?

2) we don’t have a way to have variants for
the depth shader, which is shared with the shadow
map shader, so had to use a uniform for testing
this.

3) this creates a new problem — culling needs to 
happen with the non-optimized version of the light
frustum (and it is not implemented here, so some
shadows disappear)


Change-Id: Ib753ded3f9c4753444c4a64a674bfa1ee36cb94e